Extract resolution of default output and fleet settings for agentless
