+++ x-pack/platform/plugins/shared/fleet/server/services/agents/agentless_agent.ts
@@ -30,6 +30,10 @@ import {
   AGENTLESS_GLOBAL_TAG_NAME_ORGANIZATION,
   AGENTLESS_GLOBAL_TAG_NAME_DIVISION,
   AGENTLESS_GLOBAL_TAG_NAME_TEAM,
+  DEFAULT_OUTPUT_ID,
+  SERVERLESS_DEFAULT_OUTPUT_ID,
+  DEFAULT_FLEET_SERVER_HOST_ID,
+  SERVERLESS_DEFAULT_FLEET_SERVER_HOST_ID,
 } from '../../constants';
 
 import { appContextService } from '../app_context';
@@ -56,6 +60,27 @@ interface AgentlessAgentErrorHandlingMessages {
 }
 
 class AgentlessAgentService {
+  public getDefaultSettings() {
+    const cloudSetup = appContextService.getCloud();
+    const isCloud = cloudSetup?.isCloudEnabled;
+    const isServerless = cloudSetup?.isServerlessEnabled;
+    const outputId = isServerless
+      ? SERVERLESS_DEFAULT_OUTPUT_ID
+      : isCloud
+      ? DEFAULT_OUTPUT_ID
+      : undefined;
+    const fleetServerId = isServerless
+      ? SERVERLESS_DEFAULT_FLEET_SERVER_HOST_ID
+      : isCloud
+      ? DEFAULT_FLEET_SERVER_HOST_ID
+      : undefined;
+
+    return {
+      outputId,
+      fleetServerId,
+    };
+  }
+
   public async createAgentlessAgent(
     esClient: ElasticsearchClient,
     soClient: SavedObjectsClientContract,
